api/pyproject.toml
[tool.poetry]
name = "instagram_clone_api"
version = "0.1.0"
description = ""
authors = ["g-kisenkov <85692332+g-kisenkov@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
django-environ = "^0.8.1"
django-rest-swagger = "^2.2.0"
drf-yasg = {extras = ["validation"], version = "^1.20.0"}
django-phonenumber-field = {extras = ["phonenumbers"], version = "^6.1.0"}
Pillow = "^9.0.1"
djangorestframework-simplejwt = "^5.1.0"
coverage = "^6.3.2"
react = "^4.3.0"
django-rest-framework-nested = "^0.0.1"
drf-nested-routers = "^0.93.4"
django-debug-toolbar = "^3.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
